Avoid using injector and more test to TransportClientTests
